refactor:[4-3] LottoStrategy 패키지 이동
