+++ src/main/java/lotto/domain/LottoGenerator.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import lotto.domain.lottoStrategy.LottoGenerateStrategy;
+
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
