+++ src/main/java/lotto/domain/lottoStrategy/ManualLottoStrategy.java
@@ -1,4 +1,6 @@
-package lotto.domain;
+package lotto.domain.lottoStrategy;
+
+import lotto.domain.Lotto;
 
 import java.util.List;
